Now able to call Model methods staticially.
